+++ specification/license/delete_license/DeleteLicenseRequest.ts
@@ -24,7 +24,4 @@ import { RequestBase } from '@_types/Base'
  * @since 0.0.0
  * @stability TODO
  */
-export interface Request extends RequestBase {
-  query_parameters?: {}
-  body?: {}
-}
+export interface Request extends RequestBase {}
