+++ specification/license/get_basic_license_status/GetBasicLicenseStatusRequest.ts
@@ -24,7 +24,4 @@ import { RequestBase } from '@_types/Base'
  * @since 6.3.0
  * @stability TODO
  */
-export interface Request extends RequestBase {
-  query_parameters?: {}
-  body?: {}
-}
+export interface Request extends RequestBase {}
