+++ specification/license/post_license/PostLicenseResponse.ts
@@ -17,13 +17,13 @@
  * under the License.
  */
 
-import { LicenseStatus } from '@license/get_license/LicenseStatus'
-import { LicenseAcknowledgement } from './LicenseAcknowledgement'
+import { Status } from '@license/_types/License'
+import { Acknowledgement } from './types'
 
 export class Response {
   body: {
-    acknowledge?: LicenseAcknowledgement
+    acknowledge?: Acknowledgement
     acknowledged: boolean
-    license_status: LicenseStatus
+    license_status: Status
   }
 }
