+++ specification/license/start_basic_license/StartBasicLicenseRequest.ts
@@ -28,5 +28,4 @@ export interface Request extends RequestBase {
   query_parameters?: {
     acknowledge?: boolean
   }
-  body?: {}
 }
